+++ lib/interface.ts
@@ -23,7 +23,8 @@ interface PrintSettings {
 interface PrintOptions {
     id?: string;
     name?: string;
-    path: string;
+    path?: string;
+    file?: Buffer,
     print_setting?: PrintSettings;
 }