+++ test/printer.test.js
@@ -1,4 +1,5 @@
-const {get_jobs, restart_job} = require('../dist/index')
+const fs = require('fs')
+const {get_jobs, restart_job, print_file} = require('../dist/index')
 // test('Get printer', async () => {
 //     const result = await printers()
 //     expect(result.length).not.toBe(undefined)
@@ -16,17 +17,18 @@ const {get_jobs, restart_job} = require('../dist/index')
 //     // expect(result.length).not.toBe(undefined)
 // });
 
-// test('print_pdf_sharing', async () => {
-//     const result = await print_file({
-//         id: "XFwxNzIuMzEuNjQuMjIxXEhQIEluayBUYW5rIDMxMCBzZXJpZXM=",
-//         path: "path/to/file.pdf",
-//         print_setting: {
-//             orientation: "landscape",
-//             paper: "A5"
-//         }
-//     })
-//     // expect(result.length).not.toBe(undefined)
-// });
+test('print_pdf_sharing', async () => {
+    const buffer = await fs.readFileSync("F:\\devcode\\test.pdf")
+    const result = await print_file({
+        id: "XFwxNzIuMzEuNjQuMjIxXEhQIEluayBUYW5rIDMxMCBzZXJpZXM=",
+        file: buffer,
+        print_setting: {
+            orientation: "landscape",
+            paper: "A5"
+        }
+    })
+    // expect(result.length).not.toBe(undefined)
+});
 
 // test('print_pdf_by_name', async () => {
 //     const result = await print_file({
@@ -45,7 +47,7 @@ const {get_jobs, restart_job} = require('../dist/index')
 //     console.log(result)
 // })
 
-test('restart_jobs', async () => {
-    const result = await restart_job("XFwxNzIuMzEuNjQuMjIxXEhQIEluayBUYW5rIDMxMCBzZXJpZXNfQF80Ng==");
-    // console.log(data.length)
-})
+// test('restart_jobs', async () => {
+//     const result = await restart_job("XFwxNzIuMzEuNjQuMjIxXEhQIEluayBUYW5rIDMxMCBzZXJpZXNfQF80Ng==");
+//     // console.log(data.length)
+// })